Rename `seq` module to `call`
